Fix: Shortcut consume order
